WIP: Start of work on fdconn transport.
This transport only listens, and creates connections when
the application calls setopt on the lister with NNG_OPT_FDC_FD,
to pass a file descriptor.  The FD is turned into an nng_stream,
and utilized for SP.  The protocol over the descriptor is identical
to the TCP protocol (not the IPC protocol).